# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Valkyrie REST API Specs. This service uses sessions for authentication",
        "title": "Valkyrie API",
        "contact": {},
        "license": {
            "name": "Apache 2.0"
        },
        "version": "1.0"
    },
    "host": "localhost:\u003cPORT\u003e",
    "basePath": "/api",
    "paths": {
...
        "/account/me/friends": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friends"
                ],
                "summary": "Get Current User's Friends",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Friend"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/account/me/pending": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friends"
                ],
                "summary": "Get Current User's Friend Requests",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/FriendRequest"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/account/{memberId}/friend": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friends"
                ],
                "summary": "Send Friend Request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "memberId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friends"
                ],
                "summary": "Remove Friend",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "memberId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/account/{memberId}/friend/accept": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friends"
                ],
                "summary": "Accept Friend's Request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "memberId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/account/{memberId}/friend/cancel": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friends"
                ],
                "summary": "Cancel Friend's Request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "memberId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/HttpError"
                }
            }
        },
...
        "Friend": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "isOnline": {
                    "type": "boolean"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "FriendRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "type": {
                    "description": "1: Incoming, 0: Outgoing",
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ]
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "HttpError": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "The specific error message",
                    "type": "string"
                },
                "type": {
                    "description": "The Http Response as a string",
                    "type": "string"
                }
            }
        },
...
        "SuccessResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "description": "Only returns true, not a json object",
                    "type": "boolean"
                }
            }
        },
...
    }
}